+++ src/pyLnLib/tests/calibre_test.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+
+
+from pyLnLib.calibre import calibre_test
+
+
+if __name__ == "__main__":
+    calibre_test()