+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.1.35] - 2026-08-09
+
+## Other
+
+- working on calibre.... - (Release 0.1.35) (pylnlib_commit=d944486)
+- working on calibre.... - (Release 0.1.35) (pylnlib_commit=d1fe153)
+- working on calibre.... - (Release 0.1.35) (pylnlib_commit=5c0bb2f)
+
 ## [0.1.35] - 2026-08-08
 
 ## Other